:recycle: Integrate EventConverter into EventService
