+++ src/converters/index.ts
@@ -1,3 +1,2 @@
-export * from './event_converter';
 export * from './user_task_converter';
 export * from './manual_task_converter';
