+++ src/event_service.ts
@@ -1,35 +1,47 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import {InternalServerError} from '@essential-projects/errors_ts';
 import {IEventAggregator} from '@essential-projects/event_aggregator_contracts';
 import {IIAMService, IIdentity} from '@essential-projects/iam_contracts';
 
 import {APIs, DataModels, Messages} from '@process-engine/management_api_contracts';
-import {FlowNodeInstance, IFlowNodeInstanceService, IProcessModelUseCases} from '@process-engine/persistence_api.contracts';
+import {
+  FlowNodeInstance,
+  ICorrelationService,
+  IFlowNodeInstanceService,
+  IProcessModelUseCases,
+  Model,
+} from '@process-engine/persistence_api.contracts';
+import {IProcessModelFacade, IProcessModelFacadeFactory} from '@process-engine/process_engine_contracts';
 
-import {EventConverter} from './converters/index';
 import {applyPagination} from './paginator';
+import * as ProcessModelCache from './process_model_cache';
 
 export class EventService implements APIs.IEventManagementApi {
 
+  private readonly correlationService: ICorrelationService;
   private readonly eventAggregator: IEventAggregator;
-  private readonly eventConverter: EventConverter;
   private readonly flowNodeInstanceService: IFlowNodeInstanceService;
   private readonly iamService: IIAMService;
   private readonly processModelUseCase: IProcessModelUseCases;
+  private readonly processModelFacadeFactory: IProcessModelFacadeFactory;
 
   private readonly canTriggerMessagesClaim = 'can_trigger_messages';
   private readonly canTriggerSignalsClaim = 'can_trigger_signals';
 
   constructor(
+    correlationService: ICorrelationService,
     eventAggregator: IEventAggregator,
     flowNodeInstanceService: IFlowNodeInstanceService,
     iamService: IIAMService,
+    processModelFacadeFactory: IProcessModelFacadeFactory,
     processModelUseCase: IProcessModelUseCases,
-    eventConverter: EventConverter,
   ) {
+    this.correlationService = correlationService;
     this.eventAggregator = eventAggregator;
     this.flowNodeInstanceService = flowNodeInstanceService;
     this.iamService = iamService;
+    this.processModelFacadeFactory = processModelFacadeFactory;
     this.processModelUseCase = processModelUseCase;
-    this.eventConverter = eventConverter;
   }
 
   public async getWaitingEventsForProcessModel(
@@ -43,8 +55,10 @@ export class EventService implements APIs.IEventManagementApi {
 
     const suspendedEvents = suspendedFlowNodeInstances.filter(this.isFlowNodeAnEvent);
 
-    const eventList = await this.eventConverter.convert(identity, suspendedEvents);
+    const eventList = await this.convertFlowNodeInstancesToEvents(identity, suspendedEvents);
 
+    // TODO: Remove that useless `EventList` datatype and just return an Array of Events.
+    // Goes for the other UseCases as well.
     eventList.events = applyPagination(eventList.events, offset, limit);
 
     return eventList;
@@ -72,7 +86,7 @@ export class EventService implements APIs.IEventManagementApi {
       }
     });
 
-    const eventList = await this.eventConverter.convert(identity, accessibleEvents);
+    const eventList = await this.convertFlowNodeInstancesToEvents(identity, accessibleEvents);
 
     eventList.events = applyPagination(eventList.events, offset, limit);
 
@@ -92,12 +106,12 @@ export class EventService implements APIs.IEventManagementApi {
     const suspendedEvents = suspendedFlowNodeInstances.filter((flowNode: FlowNodeInstance): boolean => {
 
       const flowNodeIsEvent = this.isFlowNodeAnEvent(flowNode);
-      const flowNodeBelongstoCorrelation = flowNode.processModelId === processModelId;
+      const flowNodeBelongsToProcessModel = flowNode.processModelId === processModelId;
 
-      return flowNodeIsEvent && flowNodeBelongstoCorrelation;
+      return flowNodeIsEvent && flowNodeBelongsToProcessModel;
     });
 
-    const eventList = await this.eventConverter.convert(identity, suspendedEvents);
+    const eventList = await this.convertFlowNodeInstancesToEvents(identity, suspendedEvents);
 
     eventList.events = applyPagination(eventList.events, offset, limit);
 
@@ -124,8 +138,88 @@ export class EventService implements APIs.IEventManagementApi {
     this.eventAggregator.publish(signalEventName, payload);
   }
 
+  private async convertFlowNodeInstancesToEvents(
+    identity: IIdentity,
+    suspendedFlowNodes: Array<FlowNodeInstance>,
+  ): Promise<DataModels.Events.EventList> {
+
+    const suspendedEvents =
+      await Promise.map(suspendedFlowNodes, async (flowNode): Promise<DataModels.Events.Event> => {
+        return this.convertToConsumerApiEvent(identity, flowNode);
+      });
+
+    const eventList: DataModels.Events.EventList = {
+      events: suspendedEvents,
+      totalCount: suspendedEvents.length,
+    };
+
+    return eventList;
+  }
+
   private isFlowNodeAnEvent(flowNodeInstance: FlowNodeInstance): boolean {
     return flowNodeInstance.eventType !== undefined;
+  }
+
+  private async convertToConsumerApiEvent(identity: IIdentity, suspendedFlowNode: FlowNodeInstance): Promise<DataModels.Events.Event> {
+
+    const processModelFacade = await this.getProcessModelForFlowNodeInstance(identity, suspendedFlowNode);
+    const flowNodeModel = processModelFacade.getFlowNodeById(suspendedFlowNode.flowNodeId);
+
+    const consumerApiEvent: DataModels.Events.Event = {
+      id: suspendedFlowNode.flowNodeId,
+      flowNodeInstanceId: suspendedFlowNode.id,
+      correlationId: suspendedFlowNode.correlationId,
+      processModelId: suspendedFlowNode.processModelId,
+      processInstanceId: suspendedFlowNode.processInstanceId,
+      eventType: <DataModels.Events.EventType> suspendedFlowNode.eventType,
+      eventName: this.getEventDefinitionFromFlowNodeModel(flowNodeModel, suspendedFlowNode.eventType),
+      bpmnType: suspendedFlowNode.flowNodeType,
+    };
+
+    return consumerApiEvent;
+  }
+
+  private async getProcessModelForFlowNodeInstance(
+    identity: IIdentity,
+    flowNodeInstance: FlowNodeInstance,
+  ): Promise<IProcessModelFacade> {
+
+    let processModel: Model.Process;
+
+    // We must store the ProcessModel for each user, to account for lane-restrictions.
+    // Some users may not be able to see some lanes that are visible to others.
+    const cacheKeyToUse = `${flowNodeInstance.processInstanceId}-${identity.userId}`;
+
+    const cacheHasMatchingEntry = ProcessModelCache.hasEntry(cacheKeyToUse);
+    if (cacheHasMatchingEntry) {
+      processModel = ProcessModelCache.get(cacheKeyToUse);
+    } else {
+      const processModelHash = await this.getProcessModelHashForProcessInstance(identity, flowNodeInstance.processInstanceId);
+      processModel = await this.processModelUseCase.getByHash(identity, flowNodeInstance.processModelId, processModelHash);
+      ProcessModelCache.add(cacheKeyToUse, processModel);
+    }
+
+    const processModelFacade = this.processModelFacadeFactory.create(processModel);
+
+    return processModelFacade;
+  }
+
+  private async getProcessModelHashForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<string> {
+    const processInstance = await this.correlationService.getByProcessInstanceId(identity, processInstanceId);
+
+    return processInstance.hash;
+  }
+
+  private getEventDefinitionFromFlowNodeModel(flowNodeModel: Model.Events.Event, eventType: string): string {
+
+    switch (eventType) {
+      case DataModels.Events.EventType.messageEvent:
+        return (flowNodeModel as any).messageEventDefinition.name;
+      case DataModels.Events.EventType.signalEvent:
+        return (flowNodeModel as any).signalEventDefinition.name;
+      default:
+        throw new InternalServerError(`${flowNodeModel.id} is not a triggerable event!`);
+    }
   }
 
 }
